fix(silk): evaluate silk_sigmoid in double and set nFramesPerPacket for stereo channels
silk_sigmoid computes 1.0 / (1.0 + exp(-x)) in double and rounds once; the
intermediate float32 roundings flipped coding_quality by one ulp at the
stereo mid/side rates. The stereo packet path now sets nFramesPerPacket
on both channels, so silk_LTP_scale_ctrl sees the packet-loss estimate
(LTP_scaleIndex was always 0 for stereo frames with FEC). The stereo flow
records per-frame LR_to_MS decisions and channel entry states for the
oracle.

+++ internal/silk/noise_shape_flp32.go
@@ -41,9 +41,10 @@ func c32(x float64) float64 { return float64(float32(x)) }
 // silkLog2FLP32 is silk_log2: (silk_float)(3.32192809488736 * log10(x)).
 func silkLog2FLP32(x float64) float64 { return f32(3.32192809488736 * math.Log10(x)) }
 
-// silkSigmoidFLP32 is silk_sigmoid: 1.0f / (1.0f + (silk_float)exp(-x)).
+// silkSigmoidFLP32 is silk_sigmoid: (silk_float)(1.0 / (1.0 + exp(-x))),
+// evaluated in double and rounded once.
 func silkSigmoidFLP32(x float64) float64 {
-	return f32(1.0 / f32(1.0+f32(math.Exp(-x))))
+	return f32(1.0 / (1.0 + math.Exp(-x)))
 }
 
 // silkWarpedAutocorrelationFLP32 ports silk_warped_autocorrelation_FLP: double
